new util to display info about revisions for one or more pages from XML input
Stub input is recommended in all cases as much faster to process than page
content input, and in particular it is required in order to display
byte counts of revisions.

This util can be used in lieu of a long query requiring a scan of the entire
page and revision tables for a wiki, to get the numbers we need to break
up the page content dump into a bunch of small reasonably-sized jobs.

Bug: T263319
Change-Id: Ia6996ee2204fb4649bf772f39b5d40bbf453241e