Script to show offsets and crcs of byte-aligned blocks of bz2 files
This lets us quickly find a block near the end of a truncated xml dump
bzip2 file where a copy of the file can be truncated, the remaining
data can be uncompressed via dumpbz2filfromoffset and everything
including the last </page> tage can be kept, via appendbz2.

Still missing: a utility to display offsets and crcs of all
blocks, even if not bit-aligned; this will be used to generate
a combined crc that can be fed to appendbz2.

Bug: T268417
Change-Id: Ibb873d0dedbc62ff298359516a53f09d7cbc74a3